🐛 Address CodeRabbit review findings on the Flint changes
All four verified against the code before fixing; each now has a regression
test that fails without its fix.

Strip raw HTML when rendering Flint Markdown. `Str::markdown` allows raw HTML
by default, and both the digest summary and topic content reach `{!! !!}`.
The summary is written from newsletters, bookmarks and fetched pages, so HTML
arriving in it is attacker-influenced. Both calls now pass
`html_input: strip` and `allow_unsafe_links: false`. Note this pattern exists
at ~20 other sites in the app (newsletter and fetch summaries among them);
those are untouched here and want their own pass.

Stop an update blanking a topic title. `sometimes|string|max:255` accepted an
empty string, and both the MCP tool and the Livewire editor send `title` on
every update, so clearing the field saved a topic with no identifier. Used
`sometimes|required` rather than the suggested `min:1`, which would still have
accepted a whitespace-only title.

Cast `diffInSeconds()` before returning it as int. Carbon 3.13.2 returns a
signed float, which `max()` propagates into an `int` return type and trips
PHP's implicit float-to-int deprecation. Fixed in TriggerFlintRoutineJob and
in TriggerFlintDigestRoutineJob, which carries the identical line — leaving
one of a matched pair fixed would only invite the same report again.

Stop the push headline doubling its full stop. `Str::before('. ')` returns the
whole line when the summary is a single sentence, so "Your evening summary."
became "Your evening summary..". Now appends a full stop only when the
sentence lacks terminal punctuation, which also covers the `!` and `?` cases
the report did not mention.

+++ app/Jobs/Flint/TriggerFlintRoutineJob.php
@@ -192,6 +192,8 @@ class TriggerFlintRoutineJob implements ShouldQueue
     {
         $endOfDay = Carbon::parse($this->localDate, $this->timezone)->endOfDay();
 
-        return max(60, now()->diffInSeconds($endOfDay, false));
+        // Carbon 3 returns a signed float here, which max() propagates into this
+        // method's int return type — cast rather than rely on implicit conversion.
+        return max(60, (int) now()->diffInSeconds($endOfDay, false));
     }
 }
